Add files via upload
These files are part of a Postman project demonstrating multi-faceted API testing carried out over a period of time using both demo API webservices sites and fake APIs generated locally in my system. This was achieved in part through my own research and reading the Postman documentations and self learning and in part through following Pavan Kumar's (SDET) YouTube tutorial channel available open source. I feel an acknowledgement and appreciation is due to him in this regard. The sole purpose of this project was to learn and also to demonstrate to potential employers my working ability to use the tool.